Working api, merged with server
