+++ README.md
@@ -4,33 +4,27 @@ A multiplayer gardening game.
 
 ## Installation
 
+### Docker
+
+Make sure you have the `docker` daemon installed and running, and `docker-compose` installed.
+
+Create a .env file using the example.env template.
+
+For development, run `docker-compose -f docker-compose.yml -f docker-compose-dev.yml build`
+
+---
+
 ### Manual
 
-For the API and game servers, you'll need `python3` and `pip`
+For the server, you'll need `python3` and `pip`
 
 For developing the React app, you'll need `create-react-app`, or at least `npm`
 
-#### API Server
+#### Server
 
-Switch to the `api/` directory
+Switch to the `server/` directory
 
-Create and activate virtual environment using virtualenv
-```bash
-$ python3 -m venv python3-virtualenv
-$ source python3-virtualenv/bin/activate
-```
-
-Use the package manager [pip](https://pip.pypa.io/en/stable/) to install all dependencies
-
-```bash
-pip install -r requirements.txt
-```
-
-#### Game Server
-
-Switch to the `socket/` directory
-
-Create and activate virtual environment using virtualenv. Make sure to use a separate one from the API server!
+Create and activate virtual environment using virtualenv.
 ```bash
 $ python3 -m venv python3-virtualenv
 $ source python3-virtualenv/bin/activate
@@ -50,30 +44,21 @@ $ npm install package.json
 
 ```
 
+## Usage
+
 ### Docker
 
-Make sure you have the `docker` daemon installed and running, and `docker-compose` installed.
+For development, run `docker-compose -f docker-compose.yml -f docker-compose-dev.yml up`
 
-Create a .env file using the example.env template.
-
-For development, run `docker-compose -f docker-compose.yml -f docker-compose-dev.yml build`
-
-## Usage
+---
 
 ### Manual
 
-#### API Server
-
-Create a .env file inside the /api directory using the example.env template.
-
-Start flask development server, using the virtualenv inside the /api directory.
-```bash
-$ export FLASK_ENV=development
-$ flask run
-```
 #### Game Server
 
-Source the virtualenv inside the /socket directory.
+Make sure you have a Postgres instance running, and configure your .env file to use it using the example.env template.
+
+Source the virtualenv.
 
 Start the server
 ```bash
@@ -83,10 +68,6 @@ uwsgi --http :5000 --gevent 1000 --http-websockets --master --wsgi-file app.py -
 #### React App
 
 Start React development server with `npm run start`
-
-### Docker
-
-For development, run `docker-compose -f docker-compose.yml -f docker-compose-dev.yml up`
 
 ## Contributing
 Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
