--- api/wsgi.py
@@ -1,5 +1,0 @@
-#This file should be created on the main folder, not in the app folder
-from app import app
-
-if __name__ == "__main__":
-	app.run()
