+++ server/Dockerfile
@@ -4,12 +4,12 @@ RUN apt-get update \
 && apt-get install gcc -y \
 && apt-get clean
 
-RUN mkdir /socket-server
-COPY requirements.txt /socket-server
-WORKDIR /socket-server
+RUN mkdir /server
+COPY requirements.txt /server
+WORKDIR /server
 RUN pip3 install -r requirements.txt
 
-COPY . /socket-server
+COPY . /server
 
 RUN chmod u+x ./entrypoint.sh
 ENTRYPOINT [ "./entrypoint.sh" ]
