+++ server/requirements.txt
@@ -13,3 +13,12 @@ uWSGI==2.0.19.1
 Werkzeug==2.0.1
 zope.event==4.5.0
 zope.interface==5.4.0
+alembic==1.6.5
+Flask-Migrate==3.0.1
+Flask-SQLAlchemy==2.5.1
+Mako==1.1.4
+psycopg2-binary==2.9.1
+python-dateutil==2.8.1
+python-editor==1.0.4
+six==1.16.0
+SQLAlchemy==1.4.19
