+++ server/src/level/skeletonGenerator.py
