--- socket/app.py
@@ -1,7 +1,0 @@
-#!/bin/env python
-from src import create_app, socketio
-
-app = create_app()
-
-if __name__ == '__main__':
-    socketio.run(app)
